test(agents): isolate exec description contracts
