+++ src/agents/bash-tools.test.ts
@@ -687,46 +687,6 @@ const runNotifyNoopCase = async ({ label, defaults, expectNotification }: Notify
   expectNotifyNoopEvents(events, expectNotification, sessionId, label);
 };
 
-describe("tool descriptions", () => {
-  it("adds automation follow-up guidance only when the scheduler is available", () => {
-    const execWithCron = createTestExecTool({ hasCronTool: true });
-    const processWithCron = createProcessTool({ hasCronTool: true });
-
-    expect(execWithCron.description).toContain(
-      "automatic completion wake when enabled and output/failure occurs; otherwise process confirms completion",
-    );
-    expect(processWithCron.description).toContain("completion without auto-wake");
-    expect(processWithCron.description).toContain("write, send-keys, submit, paste, kill");
-    expect(execWithCron.description).toContain(
-      "No sleep loops for reminders/follow-ups; use automations.",
-    );
-    expect(processWithCron.description).toContain(
-      "No polling as timer/reminder; scheduled follow-up uses automations.",
-    );
-    expect(execTool.description).not.toContain("use cron instead");
-    expect(processTool.description).not.toContain("scheduled follow-ups");
-    expect(execTool.description).toContain("otherwise process confirms completion");
-    expect(processTool.description).toContain("completion without auto-wake");
-    expect(processTool.description).toContain("write, send-keys, submit, paste, kill");
-  });
-
-  it.each(["darwin", "linux", "win32"] as const)(
-    "limits shell-quoting guidance to Unix hosts: %s",
-    (platform) => {
-      const platformSpy = vi.spyOn(process, "platform", "get").mockReturnValue(platform);
-      try {
-        expect(
-          execTool.description.includes(
-            "Quote arguments containing shell metacharacters, including URL query strings with `?` or `&`.",
-          ),
-        ).toBe(platform !== "win32");
-      } finally {
-        platformSpy.mockRestore();
-      }
-    },
-  );
-});
-
 beforeEach(() => {
   callIdCounter = 0;
   resetProcessRegistryForTests();
